:recycle: Refactor server.js and adept api.js to modifications
