+++ client/src/pages/AddTour.js
@@ -47,7 +47,7 @@ export default function AddTour() {
       const doFetch = async () => {
         try {
           const data = await getDataByID({
-            collectionName: "tasks",
+            collectionName: "tours",
             id: id,
           });
           setTask(data);
