+++ lib/serverMethods.js
@@ -35,15 +35,15 @@ async function getListByKey({ collectionName, key }) {
   return data;
 }
 
-async function insertData(collectionName, data) {
+async function insertData({ collectionName, data }) {
   return await collection(collectionName).insertOne(data);
 }
 
-async function deleteData(collectionName, id) {
+async function deleteData({ collectionName, id }) {
   await collection(collectionName).deleteOne({ _id: ObjectId(id) });
 }
 
-async function updateData(collectionName, id, data) {
+async function updateData({ collectionName, id, data }) {
   await collection(collectionName).updateOne(
     { _id: ObjectId(id) },
     { $set: data }
