+++ server.js
@@ -3,117 +3,17 @@ require("dotenv").config();
 const express = require("express");
 const path = require("path");
 const { connectToDb } = require("./lib/database");
-// const {
-//   getCollection,
-//   insertData,
-//   deleteData,
-//   updateData,
-//   getByID,
-//   getListByKey,
-// } = require("./lib/serverMethods");
 const app = express();
 const port = process.env.PORT || 3600;
 const tours = require("./routes/tours");
 const riders = require("./routes/riders");
-const customers = require("./router/customers");
+const customers = require("./routes/customers");
 
 app.use(express.json());
 
 app.use("/api/tours", tours);
 app.use("/api/riders", riders);
 app.use("/api/customers", customers);
-
-// app
-//   .route("/api/:collectionName")
-//   .get(async (req, res) => {
-//     const { name, value, sortBy, filterBy, filterValue } = req.query;
-//     const order = req.query.order === "desc" ? -1 : 1;
-//     const { collectionName } = req.params;
-//     try {
-//       const collectionData = await getCollection({
-//         collectionName,
-//         name,
-//         value,
-//         sortBy,
-//         filterBy,
-//         filterValue,
-//         order,
-//       });
-//       res.send(collectionData);
-//     } catch (e) {
-//       console.error(e);
-//       res
-//         .status(500)
-//         .send("An unexpected server error occured. Please try again later.");
-//     }
-//   })
-//   .post(async (req, res) => {
-//     const { collectionName } = req.params;
-//     try {
-//       await insertData(collectionName, req.body);
-//       res.send("New Input posted into database.");
-//     } catch (e) {
-//       console.error(e);
-//       res
-//         .status(500)
-//         .send("An unexpected server error occured. Please try again later.");
-//     }
-//   })
-//   .delete(async (req, res) => {
-//     const { id } = req.query;
-//     const { collectionName } = req.params;
-//     try {
-//       deleteData(collectionName, id);
-//       res.send("Data deleted.");
-//     } catch (e) {
-//       console.error(e);
-//       res
-//         .status(500)
-//         .send("An unexpected server error occured. Please try again later.");
-//     }
-//   })
-//   .patch(async (req, res) => {
-//     const { id } = req.query;
-//     const { collectionName } = req.params;
-//     try {
-//       updateData(collectionName, id, req.body);
-//       res.send("Data edited.");
-//     } catch (e) {
-//       console.error(e);
-//       res
-//         .status(500)
-//         .send("An unexpected server error occured. Please try again later.");
-//     }
-//   });
-
-// app.get("/api/:collectionName/:id", async (req, res) => {
-//   const { collectionName, id } = req.params;
-//   try {
-//     const data = await getByID({
-//       collectionName,
-//       id,
-//     });
-//     res.send(data);
-//   } catch (e) {
-//     console.error(e);
-//     res
-//       .status(500)
-//       .send("An unexpected server error occured. Please try again later.");
-//   }
-// });
-
-// app.get("/api/:collectionName/filter/:key", async (req, res) => {
-//   const { collectionName, key } = req.params;
-//   try {
-//     const data = await getListByKey({ collectionName, key });
-//     res.send(data);
-//   } catch (e) {
-//     console.error(e);
-//     res
-//       .status(500)
-//       .send("An unexpected server error occured. Please try again later.");
-//   }
-// });
 
 app.use(express.static(path.join(__dirname, "client/build")));
